+++ CHANGELOG.md
@@ -1822,6 +1822,26 @@ Kexa/services/addOn.service.ts was changed
 Kexa/services/analyse.service.ts was changed
 
 
+## 1.14.0-azurebenchmark.85.61e028f
+
+### Files added: 1
+
+Kexa/rules/azureBenchmarkRules.yaml was added
+
+### Files changed: 10
+
+Kexa/rules/azureBenchmarkRules.yaml was changed
+
+Kexa/emails/emails.ts was changed
+
+Kexa/models/azure/resource.models.ts was changed
+
+Kexa/rules/Kubernete.yaml was changed
+
+Kexa/services/addOn/azureGathering.service.ts was changed
+
+Kexa/services/addOn/display/azureDisplay.service.ts was changed
+
 ## 1.14.0-devestebancontinuousscan.80.7266ea8
 
 ### Files added: 2
@@ -1849,6 +1869,11 @@ Kexa/services/analyse.service.ts was changed
 Kexa/services/manageVarEnvironnement.service.ts was changed
 
 Kexa/services/updateCapability.service.ts was changed
+
+package.json was changed
+
+pnpm-lock.yaml was changed
+
 
 documentation/Documentation-Kexa.md was changed
 
@@ -1888,5 +1913,37 @@ VERSION was changed
 documentation/Documentation-Kexa.md was changed
 
 package.json was changed
+
+## 1.14.0-azurebenchmark.94.2e08853
+
+### Files added: 1
+
+Kexa/rules/azureBenchmarkRules.yaml was added
+
+### Files changed: 12
+
+Kexa/rules/azureBenchmarkRules.yaml was changed
+
+CHANGELOG.md was changed
+
+Kexa/emails/emails.ts was changed
+
+Kexa/models/azure/resource.models.ts was changed
+
+Kexa/rules/Kubernete.yaml was changed
+
+Kexa/services/addOn/azureGathering.service.ts was changed
+
+Kexa/services/addOn/display/azureDisplay.service.ts was changed
+
+Kexa/services/manageVarEnvironnement.service.ts was changed
+
+VERSION was changed
+
+capacity.json was changed
+
+package.json was changed
+
+pnpm-lock.yaml was changed